made the 'thoughts' list be in the single game page as comments, because comments don't work for some reason. also changed the header text
